Lab Background Redux
Need to redo the layout plan for the Lab page. Once done this plage should work as a tenplate for other pages and we should be good to go. Might be worth making a template page in the first place and save pieces of code around. I should do comments do I can self-reference.

I am painting with a hammer.